Cleanup, More Signatures
